complete exercise 4.21

+++ part4/bloglist/controllers/blogs.js
@@ -37,6 +37,18 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, config.SECRET)
+  if (!decodedToken) {
+    return response.status(400).json({ error: 'invalid token' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  const user = await User.findById(decodedToken.id)
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(400).json({ error: 'unauthorized request' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
